data_ingester/requirements.txt
requests
pytest
mysql-connector-python
